% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Application Server Function}
\usage{
app_server(input = NULL, output = NULL, session = NULL, volumes = NULL)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{volumes}{Named list of available volumes for file operations. If NULL,
default volumes will be created based on the operating system.}
}
\value{
Server function for use with shinyApp
}
\description{
Creates the server logic for the shinyFiles demonstration application
}
\examples{
if (interactive()) {
  server <- app_server()
  # Use with app_ui() in shinyApp()
}
}
