% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_volumes.R
\name{get_default_volumes}
\alias{get_default_volumes}
\title{Get Default Volumes Based on Operating System}
\usage{
get_default_volumes()
}
\value{
Named list of volume paths
}
\description{
Creates a list of default file system volumes appropriate for the current OS
}
\keyword{internal}
