% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_adv_feat.R
\name{mod_adv_feat_server}
\alias{mod_adv_feat_server}
\title{Advanced Features Module Server}
\usage{
mod_adv_feat_server(id, volumes)
}
\arguments{
\item{id}{Module namespace ID}

\item{volumes}{Named list of available volumes}
}
\value{
Reactive values containing advanced feature data
}
\description{
Server logic for advanced shinyFiles features
}
