% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_adv_feat.R
\name{mod_adv_feat_ui}
\alias{mod_adv_feat_ui}
\title{Advanced Features Module UI}
\usage{
mod_adv_feat_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Creates UI for demonstrating advanced shinyFiles features
}
