% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dir_browser.R
\name{mod_dir_browser_server}
\alias{mod_dir_browser_server}
\title{Directory Browser Module Server}
\usage{
mod_dir_browser_server(id, volumes)
}
\arguments{
\item{id}{Module namespace ID}

\item{volumes}{Named list of available volumes for directory operations}
}
\value{
Reactive values containing directory information
}
\description{
Server logic for directory browsing demonstrations
}
