% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dir_browser.R
\name{mod_dir_browser_ui}
\alias{mod_dir_browser_ui}
\title{Directory Browser Module UI}
\usage{
mod_dir_browser_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Creates UI for demonstrating shinyFiles directory browsing capabilities
}
