% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_select.R
\name{mod_file_select_server}
\alias{mod_file_select_server}
\title{File Selection Module Server}
\usage{
mod_file_select_server(id, volumes)
}
\arguments{
\item{id}{Module namespace ID}

\item{volumes}{Named list of available volumes for file operations}
}
\value{
Reactive values containing selected files and data
}
\description{
Server logic for file selection demonstrations
}
