% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_select.R
\name{mod_file_select_ui}
\alias{mod_file_select_ui}
\title{File Selection Module UI}
\usage{
mod_file_select_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Creates UI for demonstrating shinyFiles file selection capabilities
}
